+++ changelog.md
@@ -5,7 +5,7 @@
 * Upgraded Spring Boot from 3.0.8 to 3.1.4
 * Upgraded kotlin from 1.9.0 to 1.9.10
 * Upgraded jobrunr from 6.3.0 to 6.3.2
-* Upgraded spring data mongodb from 4.0.8 to 4.0.10
+* Upgraded spring data mongodb from 4.0.8 to 4.1.4
 * Upgraded jackson from version 2.14.3 to 2.15.2
 
 ### 0.16.6 (2023-08-15)